+++ src/features/world/index.js
@@ -4,7 +4,7 @@ import ReactTimeout         from 'react-timeout';
 
 import Map              from '../map';
 import Monsters         from '../monsters';
-import takeMonstersTurn from '../monsters/take-monsters-turn';
+import takeMonstersTurn from '../monsters/actions/take-monsters-turn';
 import Player           from '../player';
 import exploreTiles     from '../player/actions/explore-tiles';
 import generateMonsters from '../../modules/generate-monsters';
